video_dubber/requirements.txt
PySide6
faster-whisper
openai-whisper
python-dotenv
httpx
edge-tts
pydub
soundfile
pyinstaller
